+++ packages/imperative/src/auth/__tests__/BaseAuthHandler.test.ts
@@ -11,8 +11,9 @@
 
 import { BaseAuthHandler } from "../handlers/BaseAuthHandler";
 import { ICommandArguments, IHandlerParameters } from "../../../../cmd";
-import { ISession, AbstractSession, SessConstants } from "../../../../rest";
+import { Session, ISession, AbstractSession, SessConstants } from "../../../../rest";
 import { Imperative } from "../../Imperative";
+import { cloneDeep } from "lodash";
 
 class FakeAuthHandler extends BaseAuthHandler {
     public mProfileType: string = "base";
@@ -42,7 +43,7 @@ describe("BaseAuthHandler", () => {
         });
     });
 
-    it("should process login successfully", async () => {
+    it("should process login successfully - basic auth", async () => {
         const handler = new FakeAuthHandler();
         const params: IHandlerParameters = {
             response: {
@@ -73,6 +74,77 @@ describe("BaseAuthHandler", () => {
 
         expect(caughtError).toBeUndefined();
         expect(doLoginSpy).toBeCalledTimes(1);
+    });
+
+    it("should process login successfully - certificate auth", async () => {
+        const handler = new FakeAuthHandler();
+        const params: IHandlerParameters = {
+            response: {
+                console: {
+                    log: jest.fn()
+                }
+            },
+            arguments: {
+                certFile: "/fake/file/path.cert",
+                certKeyFile: "/fake/file/path.key"
+            },
+            positionals: ["auth", "login"],
+            profiles: {
+                getMeta: jest.fn(() => ({
+                    name: "fakeName"
+                }))
+            }
+        } as any;
+
+        const doLoginSpy = jest.spyOn(handler as any, "doLogin");
+        let sessionCopy;
+
+        doLoginSpy.mockImplementation((sessCfg: Session) => {
+            sessionCopy = cloneDeep(sessCfg);
+            return "fakeToken";
+        });
+        let caughtError;
+
+        try {
+            await handler.process(params);
+        } catch (error) {
+            caughtError = error;
+        }
+
+        expect(caughtError).toBeUndefined();
+        expect(doLoginSpy).toBeCalledTimes(1);
+        expect(doLoginSpy).toBeCalledWith(expect.objectContaining({
+            mISession: {
+                basePath: "",
+                cert: "/fake/file/path.cert",
+                certKey: "/fake/file/path.key",
+                hostname: "fakeHost",
+                port: 3000,
+                protocol: "https",
+                rejectUnauthorized: true,
+                secureProtocol: "SSLv23_method",
+                storeCookie: false,
+                strictSSL: true,
+                tokenType: "jwtToken",
+                type: "cert-pem"
+            }
+        }));
+        expect(sessionCopy).toEqual(expect.objectContaining({
+            mISession: {
+                basePath: "",
+                cert: "/fake/file/path.cert",
+                certKey: "/fake/file/path.key",
+                hostname: "fakeHost",
+                port: 3000,
+                protocol: "https",
+                rejectUnauthorized: true,
+                secureProtocol: "SSLv23_method",
+                storeCookie: true,
+                strictSSL: true,
+                tokenType: "jwtToken",
+                type: "cert-pem"
+            }
+        }));
     });
 
     it("should process logout successfully", async () => {