+++ CHANGELOG.md
@@ -6,6 +6,10 @@ All notable changes to the Imperative package will be documented in this file.
 
 - BugFix: Update dependencies to resolve problems with the ansi-regex package
 
+## `4.16.0`
+
+- Enhancement: Implement the ability to authenticate using client certificates in PEM format.
+
 ## `4.15.1`
 
 - Bugfix: Update js-yaml to resolve a potential security issue